✨ Feat: 전체 페이지 라우팅 + 디버깅용 /tabs
