+++ app/auth/index.tsx
@@ -1,0 +1,5 @@
+import { Redirect } from "expo-router";
+
+export default function AuthIndex() {
+  return <Redirect href="/onboarding/splash" />;
+}
